refactor: split messages.ts and consolidate duplicate utilities
- Split 1961-line messages.ts into 6 focused files (~250 lines avg)
- Consolidated duplicate createWebSearchTool function
- Extracted common error handling into handleAIResponseError helper
- Removed duplicate buildSystemPrompt function
- Fixed all TypeScript type assertions (no more any types)
- Created centralized AI tools library
- Direct exports in messages.ts (no index.ts pattern)
- Eliminated ~500+ lines of duplicate code
