+++ convex/lib/ai_tools.ts
@@ -1,0 +1,15 @@
+/**
+ * AI Tools Library
+ *
+ * Centralized location for all AI tool definitions used across the application.
+ * These tools extend AI capabilities with external services and functionalities.
+ */
+
+export { createWebSearchTool } from "../messages/tools.js"
+
+// Future tools can be added here:
+// - createImageGenerationTool
+// - createCodeExecutionTool
+// - createDatabaseQueryTool
+// - createFileSystemTool
+// etc.
